add helper class for sharing code
